Completed needed Crud Elements

+++ src/main/java/com/cjk/stackcast/services/CommentService.java
@@ -6,6 +6,8 @@ import com.cjk.stackcast.repositories.CommentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -18,18 +20,27 @@ public class CommentService {
     public Iterable<Comment> showAll(){
         return comentRopo.findAll();
     }
-    public Comment create(Comment comment){
+    public BasicComment create(BasicComment comment){
         return comentRopo.save(comment);
     }
     public Boolean deleteComment(Long commentId){
-        comentRopo.deleteById(commentId);
-        return true;
+        Comment comment = comentRopo.getOne(commentId);
+        if(comment.getCommentId().equals(commentId)){
+            comentRopo.deleteById(commentId);
+            return true;
+        }else{
+            return false;
+
+        }
     }
-    public Iterable<Comment> findByVideoId(){
-        return null;
+    public List<String> findByVideoId(Long videoId){
+        List<String> comments = new ArrayList();
+        //comentRopo.findCommentsByVideoId(videoId)
+        for(Comment comment : comentRopo.findAll()){
+            if(comment.getVideoId().equals(videoId)){
+                comments.add(comment.getComment());
+            }
+        }
+        return comments;
     }
-
-
-
-
 }
